refactor(laser)!: split phaser and r3f behind subpath entry points
The root barrel unioned modules needing mutually exclusive optional peers
— phaser + rapier, three + fiber + drei, bitecs. Every consumer imports
@kbve/laser and drags in all of them. laser is consumed as source (a
tsconfig path plus a vite alias), so its bare imports resolve relative to
packages/npm/laser, which has no node_modules. At the repo root the
hoisted tree covers that; a container's bounded install does not, so vite
substitutes an optional-peer stub and rollup dies on the named import.

That failed three times before this: bitecs ("query is not exported by
__vite-optional-peer-dep:bitecs"), then phaser and
@phaserjs/rapier-connector ("Could not resolve 'phaser' imported by
@kbve/laser"), each patched with another dedupe entry in the consumer.
drei was the fourth (#15034) and the one the dedupe list missed.

Move the renderer stacks behind entry points the ecs/mecs subpaths
already established:

  @kbve/laser/phaser  lib/phaser/* + physics/rapier
  @kbve/laser/r3f     lib/r3f/* + webgl/pom

The root barrel keeps core, spatial, tile, ecs, determ, combat, net,
auth, embed, promo, i18n — bitecs, fastnoise-lite and react only. Type-
only phaser imports stay; they are erased before resolution.

The published package gains real subpath builds. vite's lib entry was
single-entry, so ecs and mecs resolved for source consumers but did not
exist in the tarball; it is now multi-entry and package.json exports
declares all five. The root entry keeps its laser.es.js filename.

Two workarounds this removes:

  - arpg's stubLaserR3F() vite plugin, which virtual-moduled away
    lib/r3f/ because the barrel forced it into a game that never renders
    with three.
  - arpg's laser-vitest-stub.ts, a hand-written re-export of the "pure
    leaves" that node-env vitest could load. The barrel is renderer-free
    now, so vitest aliases straight to source.

BREAKING CHANGE: @kbve/laser no longer exports PhaserGame, usePhaserGame,
PlayerController, VirtualJoystick, the entity-fx/gpu-sprite-layer/
ambient-dust/object-pool/keyboard-map/arrow-projectile/bird helpers,
RAPIER, createRapierPhysics, Stage, useGameLoop, or the POM primitives
from its root. Import them from @kbve/laser/phaser or @kbve/laser/r3f.
