+++ .github/ci-dispatch-manifest.json
@@ -602,7 +602,7 @@
 		{
 			"key": "laser",
 			"package_name": "laser",
-			"version": "0.1.6",
+			"version": "0.2.0",
 			"version_toml": "packages/npm/laser/version.toml",
 			"version_source": "apps/kbve/astro-kbve/src/content/docs/project/laser.mdx",
 			"version_target": "packages/npm/laser/package.json"
