+++ apps/agones/arpg/web/tsconfig.json
@@ -16,6 +16,10 @@
 		"baseUrl": ".",
 		"paths": {
 			"@kbve/laser": ["../../../../packages/npm/laser/src/index.ts"],
+			"@kbve/laser/phaser": [
+				"../../../../packages/npm/laser/src/phaser.ts"
+			],
+			"@kbve/laser/r3f": ["../../../../packages/npm/laser/src/r3f.ts"],
 			"@kbve/observ": ["../../../../packages/npm/observ/src/index.ts"],
 			"@kbve/itemdb-data": [
 				"../../../../packages/data/codegen/generated/itemdb.json"
