+++ apps/kbve/astro-kbve/src/content/docs/project/laser.mdx
@@ -14,7 +14,7 @@ tags:
 key: laser
 pipeline: npm
 package_name: laser
-version: "0.1.6"
+version: "0.2.0"
 source_path: packages/npm/laser
 version_toml: packages/npm/laser/version.toml
 author: h0lybyte
