+++ apps/kbve/astro-kbve/tsconfig.json
@@ -22,6 +22,8 @@
 			],
 			"@kbve/droid": ["../../../packages/npm/droid/src/index.ts"],
 			"@kbve/laser": ["../../../packages/npm/laser/src/index.ts"],
+			"@kbve/laser/phaser": ["../../../packages/npm/laser/src/phaser.ts"],
+			"@kbve/laser/r3f": ["../../../packages/npm/laser/src/r3f.ts"],
 			"@kbve/chat": ["../../../packages/npm/chat/src/index.ts"],
 			"@kbve/chat/*": ["../../../packages/npm/chat/src/*"],
 			"@kbve/proto": [
