+++ apps/rareicon/astro-rareicon/tsconfig.json
@@ -14,6 +14,8 @@
 			],
 			"@kbve/droid": ["../../../packages/npm/droid/src/index.ts"],
 			"@kbve/laser": ["../../../packages/npm/laser/src/index.ts"],
+			"@kbve/laser/phaser": ["../../../packages/npm/laser/src/phaser.ts"],
+			"@kbve/laser/r3f": ["../../../packages/npm/laser/src/r3f.ts"],
 			"@kbve/proto": [
 				"../../../packages/data/codegen/generated/index.ts"
 			],
