+++ tsconfig.base.json
@@ -27,6 +27,8 @@
 			"@kbve/laser": ["packages/npm/laser/src/index.ts"],
 			"@kbve/laser/ecs": ["packages/npm/laser/src/ecs.ts"],
 			"@kbve/laser/mecs": ["packages/npm/laser/src/mecs.ts"],
+			"@kbve/laser/phaser": ["packages/npm/laser/src/phaser.ts"],
+			"@kbve/laser/r3f": ["packages/npm/laser/src/r3f.ts"],
 			"@kbve/npcdb": ["packages/data/codegen/npcdb.ts"],
 			"@kbve/spelldb": ["packages/data/codegen/spelldb.ts"],
 			"@kbve/observ": ["packages/npm/observ/src/index.ts"],
